+++ package.json
@@ -20,11 +20,11 @@
     "version": "changeset version && CI=false pnpm install"
   },
   "devDependencies": {
-    "@6edesign/cli": "workspace:^0.1.10",
-    "@changesets/cli": "^2.27.1",
-    "turbo": "^2.0.0",
-    "typescript": "^5.9.2",
-    "tsdown": "^0.13.0"
+    "@6edesign/cli": "workspace:^",
+    "@changesets/cli": "^2.29.7",
+    "tsdown": "^0.13.5",
+    "turbo": "^2.5.8",
+    "typescript": "^5.9.3"
   },
   "workspaces": [
     "packages/*"
