+++ packages/cicd/package.json
@@ -8,7 +8,7 @@
     "access": "public"
   },
   "scripts": {
-    "dev": "tsdown --copy src/templates --watch --sourcemap",
+    "dev": "tsdown --watch --sourcemap --copy src/templates src/index.ts",
     "build": "tsdown --sourcemap --copy src/templates",
     "test": "vitest run --dir test"
   },
@@ -30,7 +30,7 @@
     "@6edesign/tsconfig": "workspace:^0.0.6",
     "@vitest/coverage-v8": "^3.2.4",
     "execa": "^9.6.0",
-    "tsdown": "^0.13.0",
+    "tsdown": "^0.15.2",
     "vitest": "^1.6.0",
     "zod": "^4.0.5"
   },
