+++ packages/cicd/tsconfig.json
@@ -6,6 +6,5 @@
     "module": "esnext",
     "target": "esnext"
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "include": ["src/**/*"]
 }
