+++ packages/cli/src/commands/graph.ts
@@ -2,6 +2,7 @@ import { defineCommand } from '../utils/defineCommand';
 import { execa } from 'execa';
 import dot from '@dagrejs/graphlib-dot';
 import graphlib from '@dagrejs/graphlib';
+import z from 'zod';
 
 export async function getDependencyGraph(): Promise<graphlib.Graph> {
 	const { stdout } = await execa('pnpm', ['turbo', 'run', 'build', '--graph']);
@@ -21,14 +22,16 @@ export async function getParallelDeploymentOrder(): Promise<string[][]> {
 		inDegree[edge.w]++;
 	}
 
-	let queue = nodes.filter(node => inDegree[node] === 0);
+	let queue = nodes.filter((node) => inDegree[node] === 0);
 	const result: string[][] = [];
 
 	while (queue.length > 0) {
-		const level = queue.map(node => {
-			const match = node.match(/\[root\] (.*?)#build/);
-			return match ? match[1] : null;
-		}).filter((name): name is string => name !== null && name !== '___ROOT___');
+		const level = queue
+			.map((node) => {
+				const match = node.match(/\[root\] (.*?)#build/);
+				return match ? match[1] : null;
+			})
+			.filter((name): name is string => name !== null && name !== '___ROOT___');
 
 		if (level.length > 0) {
 			result.push(level);
@@ -49,7 +52,6 @@ export async function getParallelDeploymentOrder(): Promise<string[][]> {
 	return result;
 }
 
-
 export async function getTopologicalOrder(): Promise<string[]> {
 	const order = await getParallelDeploymentOrder();
 	return order.flat();
@@ -58,30 +60,32 @@ export async function getTopologicalOrder(): Promise<string[]> {
 export const graphCommand = defineCommand({
 	name: 'graph',
 	description: 'Get information about the workspace dependency graph',
-	subcommands: [
-		defineCommand({
-			name: 'order',
-			description: 'Get the deployment order of the workspaces',
-			handler: async () => {
+	inputs: {
+		type: {
+			schema: z.enum(['topological', 'parallel', 'json']).default('topological'),
+			description: 'The type of graph information to retrieve',
+			promptConfig: {
+				type: 'list',
+				message: 'Select the type of graph information to retrieve:',
+				choices: ['topological', 'parallel', 'json']
+			}
+		}
+	},
+	async handler(input) {
+		const actionsByType = {
+			async topological() {
 				const sorted = await getTopologicalOrder();
 				console.log(JSON.stringify(sorted, null, 2));
-			}
-		}),
-		defineCommand({
-			name: 'parallel-order',
-			description: 'Get the parallel deployment order of the workspaces',
-			handler: async () => {
+			},
+			async parallel() {
 				const sorted = await getParallelDeploymentOrder();
 				console.log(JSON.stringify(sorted, null, 2));
-			}
-		}),
-		defineCommand({
-			name: 'json',
-			description: 'Get the dependency graph in JSON format',
-			handler: async () => {
+			},
+			async json() {
 				const graph = await getDependencyGraph();
 				console.log(JSON.stringify(graphlib.json.write(graph), null, 2));
 			}
-		})
-	]
+		};
+		await actionsByType[input.type]();
+	}
 });
