+++ packages/data-circuits/package.json
@@ -8,7 +8,7 @@
     "access": "public"
   },
   "scripts": {
-    "dev": "tsdown --watch",
+    "dev": "tsdown --watch src/",
     "build": "tsdown",
     "pretest": "docker compose up -d",
     "test": "REDIS_URL=redis://localhost:6380 vitest run --dir test",
